backend/requirements.txt
# FastAPI + server
fastapi==0.116.0
uvicorn[standard]==0.35.0
pydantic[email]==2.11
greenlet==3.2.4

# Google Gemini SDK
google-generativeai==0.8.2

# Environment variables
python-dotenv==1.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Database
sqlalchemy==2.0.31
aiosqlite==0.21.0

# Data generation
faker==37.1.0

# Data analysis dependencies (for MCP functionality)
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
openpyxl>=3.1.0

# File handling
python-multipart>=0.0.6